+++ lib/ooxml_parser/common_parser/common_data/alternate_content/drawing/graphic/shape/shape_properties/docx_shape_properties/preset_geometry/shape_adjust_value_list/shape_guide.rb
@@ -1,0 +1,25 @@
+module OoxmlParser
+  # Class for describing Shape Guide
+  class ShapeGuide
+    # @return [String] name of guide
+    attr_accessor :name
+    # @return [String] shape guide formula
+    attr_accessor :formula
+
+    # Parse Relationships
+    # @param [Nokogiri::XML:Node] node with Shape Guide
+    # @return [ShapeGuide] result of parsing
+    def self.parse(node)
+      guide = ShapeGuide.new
+      node.attributes.each do |key, value|
+        case key
+        when 'name'
+          guide.name = value.value.to_sym
+        when 'fmla'
+          guide.formula = value.value
+        end
+      end
+      guide
+    end
+  end
+end